# src/types/responses/admin.responses.ts
export interface IAdminSaveAndUpdateWaiterRes{
    _id:string;
    FirstName: string;
    LastName: string;
    email:string;
    password:string;
    Desgination: string;
    createdAt: string;
    updatedAt: string;
  
}